+++ csunplugged/tests/resources/views/test_barcode_checksum_poster.py
