+++ csunplugged/tests/resources/views/test_binary_cards_small.py
