+++ csunplugged/tests/resources/views/test_binary_to_alphabet.py
