+++ csunplugged/tests/resources/views/test_job_badges.py
