+++ csunplugged/tests/resources/views/test_left_right_cards.py
