+++ csunplugged/tests/resources/views/test_parity_cards.py
