+++ csunplugged/tests/resources/views/test_sorting_network_cards.py
