+++ csunplugged/tests/resources/views/test_treasure_hunt.py
